Added the University of South Wales
The University of South Wales is a merger of the University of Glamorgan and the university of newport. As a new student, I have both a me@students.southwales.ac.uk and a me@glam.ac.uk email address so the University of Glamorgan and University of Newport files should not be removed yet (older students probably only have access to one).