xtec.txt
xtec.cat is the domain of the mail addresses from the teachers that are working for the public schools of the government of Catalonia, Spain. I'm teaching IT in a high school. My students are learning Java and Scala and we are looking to develop play2 apps with the Intellij IDE.