Community colleges and minor updates
All the community colleges in Alaska, Michigan, North Dakota, South
Dakota, and Vermont were added, according to the list provided by
Community College Review. The domain of ITT Technical Institute was
corrected to include a hyphen. Removed the domain of "edu.tf" as it is
nonexistent. Removed the domain "edu.vlerick.com" as "vlerick.com" is
present. Lastly, corrected the domain of the International
Turkmen-Turkish University.