Add new school "Gymnasium Nepomucenum Rietberg"
The domain does not match the home page of the school as we are using a different system/domain for student email addresses. Please contact the administration "Frau Milsch" or "Frau Oesterschling" (see http://www.nepomucenum-rietberg.de/kontakt/) to verify the domain gnr.wwschool.de as student email domain. Thanks in advance.