The institute of Ecology is a non profit organization devoted to research and postgraduate training on ecology, environmental and sustainability science.
It has a large reputation on environmental conservation and biodiversity knowledge generation. It is very much committed to bioinformatics for biological taxonomic records as well as of molecular structures. We do not train IT professionals per se at the moment, but we are becoming very interested in the "environmental Big data" that is starting to be available.