Create dai-labor.txt
DAI-Labor @ TU Berlin. Courses offered by us: http://www.aot.tu-berlin.de/ 
DAI-Labor is basically an institute with the AOT chair at TU Berlin. The reason we'd like to add this is because the members of the lab do have a valid university login, but no access to the mail service of TU Berlin (at least that's true for me and a few of my colleagues)