Create edmonds15.org
Edmonds School District has 4 high schools, all of which include computer science and programming classes. Mountlake Terrace High School also has a 4 year STEM program which also includes computer science classes. *@edmonds15.org is given to students in the district.

Here is a link to the district website:
https://edmonds.wednet.edu/

Here is a link saying that @edmonds15.org is given to students in the district:
https://www.edmonds.wednet.edu/cms/One.aspx?portalId=306754&pageId=489178#access