Add Faculty of Science Damanhour University
add Damanhour University which is a publice university in Egypt,
Damanhour City.
This is Damanhour University website: www.damanhour.edu.eg
Unfortunatily, We do not have a dedicated webpage for every course we
have. But, we have CS department in the faculty of science and it offers
computer science program as a major/minor, you can check it from this URL:
www.damanhour.edu.eg/scimfac/Pages/Page.aspx?id=550
I am not sure I know what could be a proof of the university recognition
of this domain, but you might consider contacting: info.mis@dmu.edu.eg