Create johak.txt
Our school is already registered with domain @hak-stjo.salzburg.at

We recently switched to this new domain. We still control the old one and for our already registered students it would be great to keep the old domain in your repo.

A new link to our curriculum - second column is the IT branch with programming throughout all 5 years. We use JetBrains products from the second year on:
https://mk0johakthat7jr5r8f.kinstacdn.com/wp-content/uploads/2020/09/stundentafel-200911.pdf

Let me know if you need anything else! THX for the great work!