Create mediaschool.txt
Mediaschool is a Campus of several schools. The IT School of "system and networks" in the campus is named IRIS (https://ecoleiris.fr/). This school provides a professional training for dev and networking in two or more years. The cursus is named SIO (Services informatiques aux organisations)(IT services for organizations). It is very common in France.