Add files via upload
The school has an iLab that includes development as a part of the curriculum. A Python class is offered for first- and second-year students. The school site that showcases the Ilab is: https://sssrome.it/flagship-programs and you can confirm that this is the institution's official email because the accrediting agencies show that this is the website of the school. you can check that on: https://councilofinternationalschools.secure.force.com/Map/exmapanythingSchoolInfo?Id=0010Y00000bHE4LQAW