Add files via upload
Adding Francis Howell School District to jet brains list of educational organizations.  This is a school district in Saint Charles Misosuri.  Adding for teachers who teach computer science so they can use jet brasins IDE to learn and practice Java to provide better instruction to out students.  To be clear, this is not used by students, only teachers for learning and practicing themselves.