add domain
We are an educational institution that started the technology course in 2022. We made the release for the students' domain, I would like the release for the teachers as well.

the school official website URL :

https://cesul.br/
the school street address, including city and country :

Brazil, Francisco Beltrão - Parana - Av. Antônio de Paiva Cantelmo, 1222 - Centro.
an URL of a page on the official website where the school offers an IT-related long-term (>1 year) course

https://cesul.br/graduacao/presencial/analise-e-desenvolvimento-de-sistemas/sobre-o-curso/
https://www.cesul.br/cesul-inicia-o-curso-de-analise-e-desenvolvimento-de-sistemas/
https://jornaldebeltrao.com.br/geral/cesul-inicia-o-curso-de-analise-e-desenvolvimento-de-sistemas/
a proof, which shows that the school recognizes the domain you are submitting as an official email domain for the students.

Each student who is enrolled gains access to the student portal which is used in the domain @alunocesul.com
https://cesul.jacad.com.br/academico/aluno-v2/login