Added Tangen domain and updated Agderskolen description.
The Aust Agder and Vest Agder county in Norway has been combined and is now called Agder.
All highschool studens in this new county are now included in this domain.

Tangen is located in Agder. The tangen domain is exclusivly for the employees of Tangen highschool.
Our students is part of hte Agderskolen domain.